Update 1.0
Good morning Good afternoon Good night.

Welcome to PokeLax

The idea is to create a website with information about the franchise we love, Pokémon! , on the site you will have the classic pokédex, pokémon information such as attacks, nature, origin and even competitive.

Update 1.0 - Description:

In the alpha version we had a .png of a pokédex , however, I took it out so that it can work on an "authentic" pokédex , we also added a .js file, to work using the PokéAPI , we refactored the index and created sobre.html (which will be updated later).